Add a vagrant file to test upgrades on Debian
The upgrade-test vagrant box installs the intelmq packages from the
stable repository, then it acitivates the unstable repositories and
upgrades the installation. Then it runs all the usual tests.
This has its ansible file in its own folder to keep the ansible
directory clean and its not part of the Makefile, because its more for
being run manually.